go-proio: consolodated non-example tests into 3 files
