+++ go-proio/event_test.go
@@ -7,6 +7,25 @@ import (
 	prolcio "github.com/decibelcooper/proio/go-proio/model/lcio"
 )
 
+func TestStrip1(t *testing.T) {
+	event := NewEvent()
+	for i := 0; i < 100; i++ {
+		event.AddEntry(
+			"Particle",
+			&eic.Particle{},
+		)
+	}
+
+	for _, ID := range event.TaggedEntries("Particle") {
+		event.RemoveEntry(ID)
+	}
+
+	nEntries := len(event.AllEntries())
+	if nEntries > 0 {
+		t.Errorf("There should be no entries, but len(event.AllEntries()) = %v out of 100", nEntries)
+	}
+}
+
 func TestTagUntag1(t *testing.T) {
 	event := NewEvent()
 	id0 := event.AddEntry(
